[UPDATE] check for config directory

+++ config/my_init.d/01_initialize.sh
@@ -1,22 +1,26 @@
 #!/bin/bash
 
 INIT="/etc/php5/initialize.sh"
-files=($(find /app/config/php56 -type f))
 
-for source in "${files[@]}" 
-do
-	pattern="\.DS_Store"
-	target=${source/\/app\/config\/php56/\/etc\/php5}
-	
-	if [[ ! $target =~ $pattern ]]; then
-		if [[ -f $target ]]; then
-			echo "    Removing \"$target\"" && rm -rf $target
+if [ -d /app/config/php56 ]
+then
+	files=($(find /app/config/php56 -type f))
+
+	for source in "${files[@]}"
+	do
+		pattern="\.DS_Store"
+		target=${source/\/app\/config\/php56/\/etc\/php5}
+
+		if [[ ! $target =~ $pattern ]]; then
+			if [[ -f $target ]]; then
+				echo "    Removing \"$target\"" && rm -rf $target
+			fi
+
+			echo "    Linking \"$source\" to \"$target\"" && mkdir -p $(dirname "${target}") && ln -s $source $target
 		fi
-		
-		echo "    Linking \"$source\" to \"$target\"" && mkdir -p $(dirname "${target}") && ln -s $source $target
-	fi
-done
-/
+	done
+fi
+
 mkdir -p /app/htdocs
 mkdir -p /app/sessions
 mkdir -p /app/logs/php56
